Move core Camera stuff into `/core/`
