+++ android/src/main/java/com/mrousavy/camera/core/CameraSession.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera
+package com.mrousavy.camera.core
 
 import android.content.Context
 import android.graphics.Point
@@ -16,6 +16,15 @@ import android.os.Build
 import android.util.Log
 import android.util.Range
 import android.util.Size
+import com.mrousavy.camera.CameraNotReadyError
+import com.mrousavy.camera.CameraQueues
+import com.mrousavy.camera.CameraView
+import com.mrousavy.camera.CaptureAbortedError
+import com.mrousavy.camera.NoRecordingInProgressError
+import com.mrousavy.camera.PhotoNotEnabledError
+import com.mrousavy.camera.RecorderError
+import com.mrousavy.camera.RecordingInProgressError
+import com.mrousavy.camera.VideoNotEnabledError
 import com.mrousavy.camera.extensions.SessionType
 import com.mrousavy.camera.extensions.capture
 import com.mrousavy.camera.extensions.createCaptureSession
@@ -30,9 +39,7 @@ import com.mrousavy.camera.parsers.QualityPrioritization
 import com.mrousavy.camera.parsers.VideoCodec
 import com.mrousavy.camera.parsers.VideoFileType
 import com.mrousavy.camera.parsers.VideoStabilizationMode
-import com.mrousavy.camera.utils.PhotoOutputSynchronizer
-import com.mrousavy.camera.utils.RecordingSession
-import com.mrousavy.camera.utils.outputs.CameraOutputs
+import com.mrousavy.camera.core.outputs.CameraOutputs
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
